Move CSVConverter into its separate file and support multiple CSVs
